feat: Add ScoreSectionInputDto for CV scoring input
feat: Introduce IObjectHasher interface for object hashing functionality

refactor: Update ISectionScoreResultRepository to use SectionType enum

refactor: Modify IJobHandler and IJobQueue to use DTOs for job handling

chore: Update project file to include Constants folder

feat: Implement SectionScoreResultService for handling section score results

fix: Rename CertificationConstraints namespace for consistency

refactor: Update SectionScoreResult entity to use SectionType enum

fix: Change TotalScore0To5 type to double in SectionScore value object

chore: Configure SectionScoreResult entity for string conversion in EF

feat: Implement SHA256ObjectHasher for consistent object hashing

chore: Update migrations to specify table names correctly

refactor: Update SectionScoreResultRepository to use SectionType enum

feat: Implement ScoreCvSectionHandler for scoring CV sections

feat: Add GetSectionScoreResultsAsync endpoint to CVController

feat: Implement ScoreCvAsync endpoint in JobsController for CV scoring

chore: Register SectionScoreResultService and IObjectHasher in DI container

refactor: Replace ScoreCvJobHandler with ScoreCvSectionHandler in job processing

+++ src/service-cv-management/ServicePerfectCV.Domain/ValueObjects/SectionScore.cs
@@ -3,7 +3,7 @@
     public sealed class SectionScore
     {
         public List<CriteriaScore> CriteriaScores { get; set; } = new();
-        public int TotalScore0To5 { get; set; } = 0;
+        public double TotalScore0To5 { get; set; } = 0;
         public double Weight0To1 { get; set; } = 0;
     }
